Begin Province Parsing

+++ diplomacygm/src/diplomacy/persistence/player.rs
@@ -2,11 +2,12 @@ use std::{collections::HashSet, sync::Arc};
 
 use super::{order::PlayerOrder, province::ProvinceInfo,};
 
+#[derive(Debug)]
 pub struct PlayerInfo {
     pub name: String,
     pub color: String,
-    pub vscc: usize,
-    pub iscc: usize,
+    pub vscc: i64,
+    pub iscc: i64,
 }
 
 pub struct Player {
@@ -16,13 +17,13 @@ pub struct Player {
 }
 
 impl PlayerInfo {
-    pub fn new(name: String, color: String, vscc: usize, iscc: usize) -> Arc<PlayerInfo> {
+    pub fn new(name: String, color: String, vscc: i64, iscc: i64) -> Arc<PlayerInfo> {
         Arc::new(PlayerInfo { name, color, vscc, iscc })
     }
 }
 
 impl Player {
-    pub fn new(info: Arc<PlayerInfo>, iscc: usize, centers: HashSet<Arc<ProvinceInfo>>) -> Player {
+    pub fn new(info: Arc<PlayerInfo>, iscc: i64, centers: HashSet<Arc<ProvinceInfo>>) -> Player {
         Player {
             info, centers, build_orders: HashSet::new()
         }
